Moved core Array to mint
