+++ src/axom/mint/mesh/internal/ConnectivityArray_typed_indirection.hpp
@@ -10,7 +10,7 @@
 
 #include "axom/core/Macros.hpp"
 #include "axom/core/Types.hpp"
-#include "axom/core/Array.hpp"
+#include "axom/mint/core/Array.hpp"
 #include "axom/mint/mesh/CellTypes.hpp"
 #include "axom/mint/config.hpp"
 #include "axom/mint/mesh/internal/ConnectivityArrayHelpers.hpp"
@@ -65,15 +65,15 @@ public:
   ConnectivityArray( IndexType ID_capacity=USE_DEFAULT,
                      IndexType value_capacity=USE_DEFAULT ) :
     m_values( nullptr ),
-    m_types( new Array< CellType >( axom::internal::ZERO, 1,
+    m_types( new Array< CellType >( axom::mint::internal::ZERO, 1,
                                     ID_capacity ) ),
-    m_offsets( new Array< IndexType >( axom::internal::ZERO,
+    m_offsets( new Array< IndexType >( axom::mint::internal::ZERO,
                                        1,
                                        m_types->capacity() + 1 ) )
   {
     IndexType new_value_capacity =
       internal::calcValueCapacity( 0, getIDCapacity(), 0, value_capacity );
-    m_values = new Array< IndexType >( axom::internal::ZERO,
+    m_values = new Array< IndexType >( axom::mint::internal::ZERO,
                                        1, new_value_capacity );
 
     m_offsets->append(0);
