+++ src/axom/mint/tests/CMakeLists.txt
@@ -38,6 +38,9 @@ set( mint_tests
      ## util tests
      mint_su2_io.cpp
      mint_util_write_vtk.cpp
+
+     ## core tests
+     mint_array.cpp
    )
 
 ## lists all axom depencies used for testing
