+++ src/axom/mint/tests/mint_mesh_coordinates.cpp
@@ -5,9 +5,9 @@
 
 #include "axom/slic/interface/slic.hpp"       // for slic macros
 
-#include "axom/core/Array.hpp"                // for axom::Array
 #include "axom/core/utilities/Utilities.hpp"  // for utilities::max()
 
+#include "axom/mint/core/Array.hpp"           // for mint::Array
 #include "axom/mint/config.hpp"               // for IndexType
 #include "axom/mint/mesh/MeshCoordinates.hpp" // for MeshCoordinates
 
@@ -1069,7 +1069,7 @@ TEST( mint_mesh_coordinates, change_resize_ratio )
 {
   constexpr int NDIMS               = 3;
   constexpr double DEFAULT_RESIZE_RATIO =
-    axom::Array< double >::DEFAULT_RESIZE_RATIO;
+    Array< double >::DEFAULT_RESIZE_RATIO;
   constexpr double NEW_RESIZE_RATIO = 2.5;
 
   MeshCoordinates mc( NDIMS );
@@ -1278,14 +1278,14 @@ TEST( mint_mesh_coordinates, sidre_push_constructor )
       EXPECT_TRUE( mesh_coords.numNodes() <= mesh_coords.capacity() );
 
       IndexType capacity = SMALL_NUM_NODES * mesh_coords.getResizeRatio() + 0.5;
-      if ( capacity < axom::Array< IndexType >::MIN_DEFAULT_CAPACITY )
+      if ( capacity < Array< IndexType >::MIN_DEFAULT_CAPACITY )
       {
-        capacity = axom::Array< IndexType >::MIN_DEFAULT_CAPACITY;
+        capacity = Array< IndexType >::MIN_DEFAULT_CAPACITY;
       }
       EXPECT_EQ( mesh_coords.capacity(), capacity );
 
       // populate the coordinates, writes to the corresponding sidre views
-      axom::Array< double > xx( dim, 1, dim );
+      Array< double > xx( dim, 1, dim );
       for ( int inode=0 ; inode < SMALL_NUM_NODES ; ++inode )
       {
         for ( int j=0 ; j < dim ; ++j )
